♻️ Refactor: split NotificationNavigation component (#78)
