+++ src/components/notification/Notification.tsx
@@ -7,12 +7,9 @@ import {
   ModalTitle,
   ModalTrigger,
 } from '@/components/common/Modal'
-import { useState, type ComponentProps } from 'react'
-import { cn } from '@/utils'
+import NotificationNavigation from '@/components/notification/NotificationNavigation'
 
 export default function Notification() {
-  const [category, setCategory] = useState<'all' | 'unReaded' | 'readed'>('all')
-
   return (
     <div className="relative">
       <Modal isOverlay={false}>
@@ -30,59 +27,10 @@ export default function Notification() {
             </button>
           </ModalHeader>
           <ModalMain className="p-0">
-            <div className="flex items-center justify-center">
-              <NotificationNavigationItem
-                isClicked={category === 'all'}
-                onClick={() => {
-                  setCategory('all')
-                }}
-              >
-                전체보기 (6)
-              </NotificationNavigationItem>
-              <NotificationNavigationItem
-                isClicked={category === 'unReaded'}
-                onClick={() => {
-                  setCategory('unReaded')
-                }}
-              >
-                읽지 않음 (3)
-              </NotificationNavigationItem>
-              <NotificationNavigationItem
-                isClicked={category === 'readed'}
-                onClick={() => {
-                  setCategory('readed')
-                }}
-              >
-                읽음 (3)
-              </NotificationNavigationItem>
-            </div>
+            <NotificationNavigation />
           </ModalMain>
         </ModalContent>
       </Modal>
     </div>
-  )
-}
-
-interface NotificationNavigationItemProps extends ComponentProps<'button'> {
-  isClicked: boolean
-}
-
-function NotificationNavigationItem({
-  children,
-  isClicked,
-  ...props
-}: NotificationNavigationItemProps) {
-  return (
-    <button
-      className={cn(
-        'flex-1 cursor-pointer py-3 text-center transition-all',
-        isClicked
-          ? 'text-primary-600 border-b-primary-500 border-b-2'
-          : 'border-0 text-gray-500'
-      )}
-      {...props}
-    >
-      {children}
-    </button>
   )
 }
